Split provider bridge state cache test module
